get_info test util function moved to test.utils
