+++ users/tests/api/mutations/user_login/test_user_login_resolver.py
@@ -1,52 +1,18 @@
 import datetime
-from typing import cast, Any
+from typing import cast
 
 import strawberry
-from django.contrib.sessions.middleware import SessionMiddleware
 from django.core.handlers.asgi import ASGIRequest
 from django.test import TransactionTestCase, AsyncRequestFactory  # type: ignore
 from django.utils import timezone
-from graphql import GraphQLResolveInfo, OperationDefinitionNode, GraphQLSchema
-from graphql.pyutils import Path
-from strawberry.django.context import StrawberryDjangoContext
-from strawberry.django.views import TemporalHttpResponse
-from strawberry.field import StrawberryField
 from strawberry.types import Info
 
 from storefront_backend.api.mutation import Mutation
-from storefront_backend.api.schema import schema
+from storefront_backend.tests.utils import get_info
 from users.api.mutations.user_login.user_login_input import UserLoginInput
 from users.api.mutations.user_login.user_login_payload import UserLoginPayload
 from users.api.mutations.user_login.user_login_resolver import user_login_resolver
 from users.models import User
-
-
-def get_info(field: StrawberryField, field_name: str, return_type: Any,
-             parent_type: Any) -> Info:
-    # Faking Request
-    async_request = AsyncRequestFactory().post("/graphql")
-    middleware = SessionMiddleware(lambda x: x)
-    middleware.process_request(async_request)
-
-    def is_awaitable(a):
-        return False
-
-    # Faking Info for resolver
-    resolver_info = GraphQLResolveInfo(
-        context=StrawberryDjangoContext(async_request, TemporalHttpResponse()),
-        root_value=None,
-        variable_values={},
-        fragments={},
-        field_name=field_name,
-        operation=OperationDefinitionNode(),
-        field_nodes=[],
-        is_awaitable=is_awaitable,
-        return_type=schema.schema_converter.from_object(return_type._type_definition),
-        parent_type=schema.schema_converter.from_object(parent_type._type_definition),
-        schema=GraphQLSchema(),
-        path=Path("", "", "")
-    )
-    return Info(_raw_info=resolver_info, _field=field)
 
 
 class UserLoginResolverTest(TransactionTestCase):
